// cc-stocks-quotes/src/app/interfaces/stock-quote.ts
export interface QuoteData {
    status: boolean;
    ask: number;
    askSize: number;
    averageDailyVolume3Month: number;
    averageDailyVolume10Day: number;
    beta: number;
    bid: number;
    bidSize: number;
    bookValue: number;
    components: [];
    currency: string;
    dividendsPerShare: number;
    earningsTimestamp: number;
    earningsTimestampEnd: number;
    earningsTimestampStart: number;
    ebitda: number;
    epsCurrentYear: number;
    epsForward: number;
    epsNextQuarter: number;
    epsTrailingTwelveMonths: number;
    esgPopulated: boolean;
    exchange: string;
    exchangeDataDelayedBy: number;
    exchangeTimezoneName: string;
    exchangeTimezoneShortName: string;
    fiftyDayAverage: number;
    fiftyDayAverageChange: number;
    fiftyDayAverageChangePercent: number;
    fiftyTwoWeekHigh: number;
    fiftyTwoWeekHighChange: number;
    fiftyTwoWeekHighChangePercent: number;
    fiftyTwoWeekLow: number;
    fiftyTwoWeekLowChange: number;
    fiftyTwoWeekLowChangePercent: number;
    fiftyTwoWeekRange: string;
    firstTradeDateMilliseconds: number;
    floatShares: number;
    forwardPE: number;
    fullExchangeName: string;
    gmtOffSetMilliseconds: number;
    heldPercentInsiders: number;
    heldPercentInstitutions: number;
    language: string;
    longName: string;
    market: string;
    marketCap: number;
    marketState: string;
    messageBoardId: string;
    pageViews: {
      midTermTrend: string;
      ongTermTrend: string;
      shortTermTrend: string;
    }
    pegRatio: number;
    postMarketChange: number;
    postMarketChangePercent: number;
    postMarketPrice: number;
    postMarketTime: number;
    priceEpsCurrentYear: number;
    priceEpsNextQuarter: number;
    priceHint: number;
    priceToBook: number;
    priceToSales: number;
    quoteSourceName: string;
    quoteType: string;
    region: string;
    regularMarketChange: number;
    regularMarketChangePercent: number;
    regularMarketDayHigh: number;
    regularMarketDayLow: number;
    regularMarketDayRange: number;
    regularMarketOpen: number;
    regularMarketPreviousClose: number;
    regularMarketPrice: number;
    regularMarketTime: number;
    regularMarketVolume: number;
    revenue: number;
    sharesOutstanding: number;
    sharesShort: number;
    sharesShortPrevMonth: number;
    shortName: string;
    shortPercentFloat: number;
    shortRatio: number;
    sourceInterval: number;
    symbol: string;
    targetPriceHigh: number;
    targetPriceLow: number;
    targetPriceMean: number;
    targetPriceMedian: number;
    totalCash: number;
    tradeable: boolean;
    trailingPE: number;
    triggerable: boolean;
    twoHundredDayAverage: number;
    twoHundredDayAverageChange: number;
    twoHundredDayAverageChangePercent: number;
};